test: consolidate driverUI tests under tests/utils
src/utils/driverUI.test.tsx predated the tests/ convention (testing.md)
and left the module with two live test files after #645 added
tests/utils/driverUI.test.tsx. Merge its driverUI cases into that file,
move its connectionIconPack cases to a new mirrored
tests/utils/connectionIconPack.test.tsx, and add coverage for the new
isUrlIcon helper. No test cases were dropped.
