+++ tests/utils/driverUI.test.tsx
@@ -1,8 +1,37 @@
-import { describe, it, expect } from "vitest";
-import { render } from "@testing-library/react";
-import { getDriverIcon, getConnectionIcon } from "../../src/utils/driverUI";
+import { describe, it, expect, vi } from "vitest";
+import { render, screen, waitFor } from "@testing-library/react";
+import {
+  getConnectionAccent,
+  getConnectionIcon,
+  getDriverIcon,
+  isUrlIcon,
+} from "../../src/utils/driverUI";
 import type { PluginManifest } from "../../src/types/plugins";
 import type { SavedConnection } from "../../src/contexts/DatabaseContext";
+
+// Avoid loading the lazy ConnectionIconImage during tests (it pulls Tauri APIs that aren't available in vitest)
+vi.mock("../../src/components/ConnectionIconImage", () => ({
+  ConnectionIconImage: (props: { path: string; size: number }) =>
+    <img data-testid="conn-icon-image" alt="" src={`mock://${props.path}`} width={props.size} height={props.size} />,
+}));
+
+describe("isUrlIcon", () => {
+  it("matches http(s) URLs and data: URIs regardless of scheme case", () => {
+    expect(isUrlIcon("http://example.com/icon.svg")).toBe(true);
+    expect(isUrlIcon("https://example.com/icon.svg")).toBe(true);
+    expect(isUrlIcon("HTTPS://example.com/icon.svg")).toBe(true);
+    expect(isUrlIcon("data:image/png;base64,iVBORw0KGgo=")).toBe(true);
+    expect(isUrlIcon("DATA:image/png;base64,iVBORw0KGgo=")).toBe(true);
+  });
+
+  it("rejects built-in lookup keys and non-URL strings", () => {
+    expect(isUrlIcon("postgres")).toBe(false);
+    // "database" starts with "data" but is not a data: URI
+    expect(isUrlIcon("database")).toBe(false);
+    expect(isUrlIcon("ftp://example.com/icon.svg")).toBe(false);
+    expect(isUrlIcon("")).toBe(false);
+  });
+});
 
 /**
  * getDriverIcon's priority is documented as: manifest-supplied URL/data:
@@ -63,6 +92,21 @@ describe("getDriverIcon", () => {
   });
 });
 
+describe("getConnectionAccent", () => {
+  const manifest = { id: "mysql", color: "#0000ff", icon: "database" } as unknown as PluginManifest;
+
+  it("uses override when present", () => {
+    const c = { appearance: { accentColor: "#ff0000" } } as SavedConnection;
+    expect(getConnectionAccent(c, manifest)).toBe("#ff0000");
+  });
+  it("falls back to manifest color when override missing", () => {
+    expect(getConnectionAccent({} as SavedConnection, manifest)).toBe("#0000ff");
+  });
+  it("falls back to grey when both missing", () => {
+    expect(getConnectionAccent(null, null)).toBe("#64748b");
+  });
+});
+
 describe("getConnectionIcon", () => {
   const manifest = (icon?: string): PluginManifest => ({ icon } as PluginManifest);
   const connection = (
@@ -100,5 +144,28 @@ describe("getConnectionIcon", () => {
     const { getByTestId } = renderConnectionIcon(connection(undefined), "postgres");
     expect(getByTestId("wrap").querySelector("svg")).not.toBeNull();
     expect(getByTestId("wrap").querySelector("img")).toBeNull();
+  });
+
+  it("renders emoji when override is emoji", () => {
+    const c = { id: "1", appearance: { icon: { type: "emoji", value: "🐘" } } } as SavedConnection;
+    render(<>{getConnectionIcon(c, manifest("database"), 16)}</>);
+    expect(screen.getByText("🐘")).toBeInTheDocument();
+  });
+
+  it("falls back to manifest icon when override missing", () => {
+    // Smoke test: should not throw; icons are mocked to null in test env
+    expect(() => render(<>{getConnectionIcon({ id: "1" } as SavedConnection, manifest("database"), 16)}</>)).not.toThrow();
+  });
+
+  it("renders the mocked image component for image overrides", async () => {
+    const c = { id: "1", appearance: { icon: { type: "image", path: "connection-icons/foo.png" } } } as SavedConnection;
+    render(<>{getConnectionIcon(c, manifest("database"), 16)}</>);
+    await waitFor(() => expect(screen.getByTestId("conn-icon-image")).toBeInTheDocument());
+  });
+
+  it("falls back to manifest when pack id is unknown", () => {
+    const c = { id: "1", appearance: { icon: { type: "pack", id: "this-icon-does-not-exist-xyz" } } } as SavedConnection;
+    // Smoke test: should not throw; icons are mocked to null in test env
+    expect(() => render(<>{getConnectionIcon(c, manifest("database"), 16)}</>)).not.toThrow();
   });
 });
